feat #28084: save last composer name as hint and support autofill on Tab/Enter
When entering a composer name, the last used name is now stored and shown as a hint in the input field. If the field is empty, pressing Tab or Enter will automatically fill it with the previous composer’s name, reducing repetitive typing. Users can still type a new name at any time without needing to delete the hint explicitly. (The autofill on Tab/Enter not yet working)